backend/requirements.txt
uagents
openai
langchain
faiss-cpu==1.7.4
PyYAML
llama-index
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
pydantic==2.6.1
transformers==4.37.2
torch==2.2.0
python-dotenv==1.0.1
sentence-transformers==2.5.1
langchain==0.1.5
langchain-community==0.0.16
chromadb==0.4.22
python-multipart==0.0.6
